Default to user writable directory in config
When a user installs from a package manager ([MacPorts], [Homebrew],
`yum`, etc.), the `todo.sh` script is installed into a restricted
folder.  This means that the result of `$(basename "$0")` is a directory
where the user cannot write without elevated permissions.  This
generates additional noise when the user first runs `todo.sh -h`.  An
alternative is to default to the user's personal directory which is
known to exist and will be writable by the user.

[MacPorts]: https://www.macports.org
[Homebrew]: http://brew.sh