Refactoring: Replace serial checks for TODOTXT_CFG_FILE and TODO_ACTIONS_DIR with loops
Two fallbacks (like for the actions dir) may still be fine, but we're now supporting so many config locations that a loop is much shorter, and the various locations are much easier to see.
For consistency, also apply this to the actions dir lookup, although it's less of a problem there.