TESTSFIX: t0000-config.sh may unintentionally pick up existing user config in $XDG_CONFIG_HOME/todo/config
This corner case was missed when the XDG alternative config location was introduced, likely because the variable isn't set by default on Ubuntu or Mac OS, and none of the developers had a config there.
HOME is already redirected to the current working directory; as XDG_CONFIG_HOME is just a config alternative that has a fallback to $HOME/.config, simply unset it.

Fixes #347