ENH: Allow opt-in to sed emulation for add-ons via TODOTXT_SED_EXPORT_FOR_ADDONS flag
Most add-ons likely copied the "sed -i.bak ..." from todo.sh, but we can't be sure. Therefore, just document the issue in the configuration and offer a simple switch to enable it.
As most users are either on Linux or Windows/Cygwin where this isn't an issue, the rather technical language only applies to users of BSD or MacOS users who don't want to install GNU sed.