Robustness: Ignore failure to remove $(DESTDIR)$(datarootdir)
Unlike the other directories, this is a shared one; if there are other (locally installed) completions, it's not empty and cannot be removed. But if we're the only / last locally installed completion, it's nice to clean it up.
Suppress errors and command failure so that "make uninstall" does not fail due to this inconsequential error.

Fixes #451